// src/interfaces/response.tsx
import {ImagePicker} from './main';
// Generated by https://quicktype.io

export interface AuthResponse {
  token?: string;
  error?: string;
}

export interface SaveLessonResponse {
  lesson: Lesson;
  error?: string;
}

export interface Lesson {
  lesson: string;
  nrc: number;
  teacher: string;
  schedlue: Schedlue[];
  _id: string;
}

export interface Schedlue {
  day: string;
  hours: string;
}

export interface DeleteLessonResponse {
  error?: string;
  msg?: string;
}

export interface EditLessonResponse {
  error?: string;
  lesson: Lesson;
}

export interface TaskResponse {
  tasks: Task[];
  error?: string;
}

export interface Task {
  _id: string;
  title: string;
  body: string;
  complete: boolean;
  images: ImagePicker[];
  dayLimit: null | string;
  lesson: string;
}

export interface TaskResponseApi {
  task: Task;
  error?: string;
}

export interface ResponseTask {
  msg?: string;
  error?: string;
}

export interface ImagesResponse {
  images: UriImage[];
}

export interface UriImage {
  uri: string;
}

export interface NotesResponse {
  note: Note;
  error?: string;
}

export interface Note {
  title: string;
  body: string;
  images: ImagePicker[];
  _id: string;
  createdAt: string;
  updatedAt: string;
}

export interface Image {
  url: string;
}

export interface ActiveModal {
  task?: Task;
  note?: Note;
}

export interface UserResponse {
  user: User;
  error?: string;
}

export interface User {
  _id: string;
  username: string;
  email: string;
  image: string;
}